// src/api/model/addressAnalysis.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * compliance-queries-api
 * The ward's compliance queires endpoints
 * OpenAPI spec version: 1.0
 */
import type { Exposure } from './exposure';

export interface AddressAnalysis {
  address: string;
  blockchain: string;
  incomingDirectExposure: Exposure;
  incomingIndirectExposure: Exposure;
  labels: string[];
  outgoingDirectExposure: Exposure;
  outgoingIndirectExposure: Exposure;
  risk: number;
  timestampScanned: string;
}